added riding offsets for flygon fixed moving and standing offset setup

+++ minecraft/pixelmon/entities/pixelmon/stats/RidingOffsets.java
@@ -17,7 +17,7 @@ public class RidingOffsets {
 		standing.z = (float)z;
 	}
 
-	public void setMovingOffsets(float x, float y, float z) {
+	public void setMovingOffsets(double x, double y, double z) {
 		moving.x = (float)x;
 		moving.y = (float)y;
 		moving.z = (float)z;
